test(work): split qualified core provenance coverage
