+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.1",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.1",
-    "sourceSha": "f38d5a4844b38f57aa59f569f7cbfd4815f8990d"
+    "sourceSha": "df951dfbe96500d6bfba84db4b8d974fcb2bb68c"
   },
   "claims": [
     {
